+++ src/apple/apple-wizard.ts
@@ -13,6 +13,7 @@ import * as bash from '../utils/bash';
 import { WizardOptions } from '../utils/types';
 import * as Sentry from '@sentry/node';
 import { traceStep, withTelemetry } from '../telemetry';
+import * as cocoapod from './cocoapod';
 import * as fastlane from "./fastlane"
 
 const xcode = require('xcode');
@@ -103,10 +104,20 @@ async function runAppleWizardWithTelementry(
     options.url,
   );
 
+  const hasCocoa = cocoapod.usesCocoaPod(projectDir);
+
+  if (hasCocoa) {
+    const podAdded = await traceStep('Add CocoaPods reference', () => cocoapod.addCocoaPods(projectDir));
+    if (!podAdded) {
+      clack.log.warn("Could not add Sentry pod to your Podfile. You'll have to add it manually.\nPlease follow the instructions at https://docs.sentry.io/platforms/apple/guides/ios/#install");
+    }
+  }
+
   traceStep('Update Xcode project', () => {
-    xcManager.updateXcodeProject(pbxproj, project, apiKey, true, true);
+    xcManager.updateXcodeProject(pbxproj, project, apiKey, !hasCocoa, true);
   });
 
+  Sentry.setTag('package-manager', hasCocoa ? "cocoapods" : "SPM");
   const projSource = path.join(
     projectDir,
     xcodeProjFile.replace('.xcodeproj', ''),