+++ src/utils/bash.ts
@@ -38,7 +38,20 @@ export async function installSentryCLI(): Promise<void> {
     await httpAsync;
 }
 
-export function execute(command: string): string {
+export function executeSync(command: string): string {
     const output = child_process.execSync(command);
     return output.toString();
+}
+
+export function execute(command: string): Promise<string> {
+    return new Promise((resolve, reject) => {
+        child_process.exec(command, (error, stdout, _) => {
+            if (error) {
+                reject(error);
+                return;
+            }
+
+            resolve(stdout);
+        });
+    });
 }