+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-feat(apple): Add Fastlane detector for iOS wizard (#356)
+- feat(apple): iOS wizard has support for cocoapods (#350)
+- feat(apple): Add Fastlane detector for iOS wizard (#356)
 
 ## 3.7.1